+++ pyisim/rest.py
@@ -481,3 +481,17 @@ class ISIMClient:
         ous = self.s.get(url, params=params)
 
         return ous.json()
+
+    def lookup_person_dn(self,dn,attributes,embedded):
+        url = self.__addr + f"/itim/rest/people/person"
+
+        params = {
+            "attributes": attributes,
+            "embedded": embedded,
+            "distinguishedName;x-property":dn
+        }
+        data = urlencode(params, quote_via=urllib.parse.quote)
+
+        person = self.s.get(url, params=data)
+
+        return person.json()